Replace CrudRepository with custom BaseRepository
CrudRepository doesn't work nicely with Groovy due to overuse of
overloaded methods with generic types. Groovy handles generics a bit
differently than Java and these methods are quite ambiguous here. This
problem can be solved, but I found it cleaner and more reliable to not
use it and rather define custom interface (I'm not using Repository
support as intended anyway).