Release version 0.5.0
Wait a moment... the last version was 1.0.0.M2, wan't it? Well, this is
lame, but the decision to start with 1.0.0-SNAPSHOT was simply wrong.
Very wrong. Since there are still some core features missing, I don't
think that it can be marked as 1.0.0. But the current state is
unsustainable, I must start numbering stable, yet not feuture complete,
releases. So that's it.